move components where its used
