+++ src/pages/pokemon-details/stat-card.component.tsx
